At #444 Add Static graph Shared Component and Bar Chart
